fixed some bugs added some more
it kinda works but some wierd issues happen where sometimes it returns the wrong result or it randomly segfaults (the resulting program) which seems to be happening at random, it seems to be connected with user defined local variables and or cond. ive tracked on segfault down where after the RET it "returns" to the location of a string pointer.
something is really wrong here